Update Experimentation
So the gauge mechanic is working now. Its values can be easily accessed for other programs in the project. Each gauge has a Gauge_FIll script and a corresponding Interaction_Manager_GaugeInput script on the interaction manager. This was done for generic programming. and so that the feature can be easily changed in future.